Renamed model to unet_model
